requirements.txt
PyGithub
texttable
slackclient
